script/Windows/javbusapi/requirements.txt
python-telegram-bot[job-queue]>=20.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
